Add .gitignore file for Visual Studio Code and other platforms and libraries
This commit adds a .gitignore file generated using the '.gitignore Generator' for Visual Studio Code. It includes ignore rules for various platforms, libraries, archives, compressed archives, and more. The .gitignore file was created based on a template from the website 'toptal.com', and it has been edited to suit the needs of the project. The file was added to the root directory of the project.